+++ test/support/exchange_rate_callback_module.ex
@@ -2,15 +2,16 @@ defmodule Money.ExchangeRates.CallbackTest do
   @behaviour Money.ExchangeRates.Callback
 
   def init do
+    :ok
   end
 
   def latest_rates_retrieved(_rates, _retrieved_at) do
-    IO.puts("Latest Rates Retrieved")
+    Application.put_env(:ex_money, :test, "Latest Rates Retrieved")
     :ok
   end
 
   def historic_rates_retrieved(_rates, _date) do
-    IO.puts("Historic Rates Retrieved")
+    Application.put_env(:ex_money, :test, "Historic Rates Retrieved")
     :ok
   end
 end